# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "go-gin-backend",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/transaction": {
            "get": {
                "description": "Retrieve all transaction records",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "Retrieve Transactions of all Users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer eyJhb...",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 100,
                        "description": "maximum records per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "page of records, starts from 0",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Transaction"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPError"
                        }
                    }
                }
            }
        },
        "/transaction/{transaction_id}": {
            "put": {
                "description": "Update a Specific Transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "Update Transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer eyJhb...",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "transaction's id",
                        "name": "transaction_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Transaction"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPError"
                        }
                    }
                }
            }
        },
        "/transaction/{user_id}": {
            "get": {
                "description": "Retrieve transaction records of a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "Retrieve Transactions of User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer eyJhb...",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "user's id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 100,
                        "description": "maximum records per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "page of records, starts from 0",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Transaction"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "Create transaction records",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "Create Transactions for User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer eyJhb...",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "user's id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "transactions",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TransactionList"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Transaction"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Campaign": {
            "type": "object",
            "properties": {
                "accepted_mccs": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "bonus_rates": {
                    "type": "number",
                    "example": 8.8
                },
                "campaign_id": {
                    "type": "string",
                    "example": "cmp00001"
                },
                "card_type": {
                    "type": "string",
                    "example": "super_miles_card"
                },
                "description": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "merchant": {
                    "type": "string",
                    "example": "7-11"
                },
                "min_spend": {
                    "type": "number",
                    "example": 100
                },
                "start_date": {
                    "type": "string"
                }
            }
        },
...
        "models.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        },
        "models.Transaction": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 20.1
                },
                "campaign": {
                    "$ref": "#/definitions/models.Campaign"
                },
                "card_id": {
                    "type": "string",
                    "example": "4111222233334444"
                },
                "card_pan": {
                    "type": "string",
                    "example": "xyz"
                },
                "card_type": {
                    "type": "string",
                    "example": "super_miles_card"
                },
                "cashback": {
                    "type": "number",
                    "example": 100
                },
                "currency": {
                    "type": "string",
                    "example": "USD"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "is_deleted": {
                    "type": "boolean",
                    "example": false
                },
                "mcc": {
                    "type": "string",
                    "example": "5311"
                },
                "merchant": {
                    "type": "string",
                    "example": "7-11"
                },
                "miles": {
                    "type": "number",
                    "example": 100
                },
                "points": {
                    "type": "number",
                    "example": 100
                },
                "transaction_date": {
                    "type": "string",
                    "example": "yyyy-mm-dd hh:mm:ss"
                },
                "transaction_id": {
                    "type": "string",
                    "example": "txn00001"
                },
                "user_id": {
                    "type": "string",
                    "example": "6400a..."
                }
            }
        },
        "models.TransactionList": {
            "type": "object",
            "properties": {
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Transaction"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}